Support for NPC level integer values.

+++ L2J_Mobius_8.0_Homunculus/java/org/l2jmobius/gameserver/data/xml/NpcData.java
@@ -102,7 +102,7 @@ public class NpcData implements IXmlReader
 						List<DropHolder> dropLists = null;
 						set.set("id", npcId);
 						set.set("displayId", parseInteger(attrs, "displayId"));
-						set.set("level", parseByte(attrs, "level"));
+						set.set("level", parseInteger(attrs, "level"));
 						set.set("type", parseString(attrs, "type"));
 						set.set("name", parseString(attrs, "name"));
 						set.set("usingServerSideName", parseBoolean(attrs, "usingServerSideName"));
